YouAssist/requirements.txt
python-dotenv
langchain
openai
youtube-transcript-api
faiss-cpu
streamlit
langchain-community